skew_consistency: Add consistency check while ADJFREQ is being tweaked
This test checks that we don't see any consistency issues when we're
slowing down and speeding up the frequency asynchronously.

This is different from the change_skew test, which validates we don't
see problems at various fixed frequency adjustment points. This test
keeps the freq in flux while we watch for issues.

Signed-off-by: John Stultz <john.stultz@linaro.org>